refactor(sales): split god-component SalesPrintDashboard (827→435)
Quebra a Impressão de Pedidos em módulos sob src/components/sales/print/.
Toda a busca de dados (7 useQuery), as derivações (useMemo) e os handlers
permanecem no pai; helpers de impressão viram funções puras testáveis.

- types.ts: tipos + COMPANY_LABELS/COMPANY_COLORS + getPeriod + EnrichedOrder.
- buildPrintHtml.ts: buildPrintData + escapeHtml + buildSingleOrderHtml +
  buildPrintDocument (template do documento extraído de printSelected) — puro.
- OrderGroup.tsx: grupo de pedidos por período (controlado).
- PrintFilters.tsx: Card de filtros (data/período/empresas) — controlado.
- +16 testes (4 arquivos): helpers de HTML (parcelas, escape, recibo OBEN,
  documento), getPeriod, render e cliques de OrderGroup/PrintFilters.

tsc --noEmit limpo; lint 0 problemas.
